Index

UnitfulCosmo.mpc

UnitfulCosmo.planck

UnitfulCosmo.unmpc

UnitfulCosmo.unplanck

UnitfulCosmo.mpc — Method

mpc(q; base=u"Mpc")

Convert q to Mpc units based on the units specified by base. If base is unspecified, mpc will default to Mpc. You could choose some other length unit if you wanted.

UnitfulCosmo.planck — Method

planck(q)

Strip units by inserting whatever ħ, c, G and kb are required to make the conversion work.

UnitfulCosmo.unmpc — Method

unmpc(targetUnit, q)

Convert a quantity q to units specified by targetUnits automatically inserting whatever factors of ħ, c, ϵ₀ and kb are required to make the conversion work.

UnitfulCosmo.unplanck — Method

unplanck(targetUnit, q)

Convert a quantity q to units specified by targetUnits automatically inserting whatever factors of ħ, c, G and kb are required to make the conversion work.